Add `login` command
This commit adds the `login` command for the
user to log in to the server address passed.

This will enable users to login to multiple
registeries and specific which set of credentials
to use for other subcommands via cli or env var
else it uses the current credential set in the
config file.

This commit also adds the utils functions that
will be used by successive commits.

Signed-off-by: Akshat <akshat25iiit@gmail.com>