feat(project robot): Added function to check whether a robot with the same already exists.
To do so the robot handler now has a function that can check for existance of a robot by name. It searches for all robots in the target project and filter whether the to be created robot's name already exists there. To that end it queries the project name and robot account name prefix from the instance configuration. This is necessary since the q=name=<targetName> returns ambiguous results

Signed-off-by: Patrick Eschenbach <patrickeschenbach96@gmail.com>